int64 explicit cast

+++ ChaosMetadataService/QueryDataMsgPSConsumer.cpp
@@ -66,7 +66,7 @@ void QueryDataMsgPSConsumer::messageHandler(chaos::common::message::ele_t& data)
     chaos::common::data::CDataWrapper* cd=data.cd.get();
 
     if (cd&&cd->hasKey(DataPackCommonKey::DPCK_DATASET_TYPE) && cd->hasKey(NodeDefinitionKey::NODE_UNIQUE_ID)) {
-      int64_t now = TimingUtil::getTimeStamp();
+      int64_t now = (int64_t)TimingUtil::getTimeStamp();
 
       int pktype = cd->getInt32Value(DataPackCommonKey::DPCK_DATASET_TYPE);
 
@@ -85,7 +85,7 @@ void QueryDataMsgPSConsumer::messageHandler(chaos::common::message::ele_t& data)
         if (cd->hasKey(MetadataServerLoggingDefinitionKeyRPC::PARAM_NODE_LOGGING_LOG_TIMESTAMP)) {
           ts  = cd->getInt64Value(MetadataServerLoggingDefinitionKeyRPC::PARAM_NODE_LOGGING_LOG_TIMESTAMP);
           lat = now - ts;
-          if (lat > chaos::common::constants::SkipDatasetOlderThan) {
+          if (lat > (int64_t)chaos::common::constants::SkipDatasetOlderThan) {
             ERR <<  data.key << " log too old: " << lat/1000.0 << " s, skipping...";
             return;
           }
